% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_advice.R
\name{get_advice}
\alias{get_advice}
\alias{get_advice.MungrCleaner}
\title{Get Advice from MungrCleaner}
\usage{
get_advice(x, ...)

\method{get_advice}{MungrCleaner}(x, na_threshold = 0.5, factor_card_perc = 0.7, string_card_perc = 0.1, ...)
}
\arguments{
\item{x}{An object containing data to be processed (e.g., a \code{MungrCleaner} object).}

\item{...}{Additional arguments (not used).}

\item{na_threshold}{Numeric. The proportion of missing NA values
above which a column is flagged as CRITICAL, suggesting it should be
dropped. Default value is 0.5.}

\item{factor_card_perc}{Numeric. The ratio of unique values to total rows
above which a column (currently character/factor) is flagged as having
High Cardinality, suggesting it should be treated as an ID or free text.
Default value is 0.7.}

\item{string_card_perc}{Numeric. The ratio of unique values to total rows
below which a column (currently character/factor) is flagged as having
Low Cardinality, suggesting it should be converted to a factor.
Default is 0.1.}
}
\value{
The function is primarily called for its side effect of printing advice.
It invisibly returns the input object \code{x}.

Invisibly returns the input object \code{x} with no modifications
to the data or log.
}
\description{
This generic function analyzes a MungrCleaner object data frame and identifies common data quality issues
(missing values, empty strings, case inconsistency, and inappropriate
cardinality). It prints suggested cleaning actions to the console.
}
\details{
The function checks for the following issues for each column: missing data serverity,
empty strings, case sensitiviy, high cardinality (flags potential ID/free-text columns),
low cardinality (flags potential factor/categorical columns), and zero variance
}
\examples{
# Example data demonstrating various issues
messy_df <- data.frame(
  ID = c(1, 2, 3, 4, 5, 6),
  Category = c("A", "a", "B", "C", "C", NA),
  Notes = c(letters[1:5], ""), # Example of empty string
  Useless = rep(10, 6)
)

cleaner <- MungrCleaner(data_messy)

# Get default advice
get_advice(cleaner)

# Get advice with a stricter missing threshold (e.g., 20\%)
get_advice(cleaner, na_threshold = 0.2)
}
