% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MungCleanr.R
\name{MungrCleaner}
\alias{MungrCleaner}
\title{MungrCleaner S3 Class}
\usage{
MungrCleaner(df)
}
\arguments{
\item{df}{A data frame or object that can be coerced to a data frame intended for cleaning.}
}
\value{
An object of the MungrCleaner class, which includes two attributes: data and log.
These attributes can be accessed using the object name and $.
}
}
\description{
The constructor for the S3 class. It initializes (1) a data cleaning pipeline object along with the original dataframe
and (2) a history log attribute.
The input data is converted to a data frame for consistent handling
with tidyverse functions.
}
\examples{
# Create a simple dataframe
data <- data.frame(ID = 1:3, Value = c(10, 20, 30))

# Initialize the MungrCleaner object
cleaner <- MungrCleaner(data)

# Check the class
class(cleaner)

}
