% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MungCleanr.R
\name{print.MungrCleaner}
\alias{print.MungrCleaner}
\title{Print Method for MungrCleaner Objects}
\usage{
\method{print}{MungrCleaner}(x, ...)
}
\arguments{
\item{x}{An object of MungrCleaner.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the input object. Prints history log in console.
}
\description{
Prints the cleaning history log of the MungrCleaner object.
}
